Factor registration methods into separate class
